pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-misuse-triage"
version = "1.0.0"
description = "A reinforcement learning environment for triage and classification of AI misuse."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.1.2",
    "pytest>=7.4.3"
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["openenv_misuse_triage", "server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]